web/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
anyio==3.7.1
argcomplete==3.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Automat==0.8.0
backoff==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.1.0
BingImageCreator==0.5.0
blinker==1.4
boto3==1.28.28
botocore==1.31.28
bs4==0.0.1
celery==5.3.1
censys==2.2.4
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloud-enum @ file:///home/ubuntu/Tools/cloud_enum
cloud-init==23.2.2
cmseek @ file:///home/ubuntu/Tools/CMSeeK
colorama==0.4.6
colorclass==2.2.2
colored==1.4.4
colorlog==6.7.0
command-not-found==0.3
commix @ file:///home/ubuntu/Tools/commix
configobj==5.0.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==2.8
dank==0.0.23
DateTime==5.2
datrie==0.8.2
dbus-python==1.2.16
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.7
distro==1.4.0
distro-info==0.23+ubuntu1.1
Django==4.2.4
django-admin-view-permission==1.9
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-environ==0.10.0
django-tenant-schemas==1.12.0
django-tenants==3.5.0
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.0.0
DNSValidator @ file:///home/ubuntu/Tools/dnsvalidator
ec2-hibinit-agent==1.0.0
EdgeGPT==0.13.2
editdistance==0.6.2
emailfinder==0.3.0b0
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fake-useragent==1.2.1
favicon==0.7.0
filelock==3.12.2
frozenlist==1.4.0
future==0.18.3
ghauri @ file:///home/ubuntu/Tools/ghauri
google==3.0.0
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hibagent==1.0.1
html5lib==1.1
httpcore==0.17.3
httplib2==0.14.0
httpx==0.24.1
hyperlink==19.0.0
idna==3.4
importlib-metadata==1.5.0
importlib-resources==6.0.1
incremental==16.10.1
Interlace @ file:///home/ubuntu/Tools/interlace
ipaddress==1.0.23
ipwhois==1.2.0
Jinja2==2.10.1
jmespath==1.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==1.1.0
mdurl==0.1.2
metafinder==1.2
mmh3==4.0.1
more-itertools==4.2.0
multidict==6.0.4
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
openpyxl==3.1.2
ordered-set==4.1.0
packaging==23.1
pexpect==4.6.0
pikepdf==8.4.0
Pillow==10.0.0
platformdirs==3.10.0
pluginbase==1.0.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybind11==2.11.1
pyfiglet==0.8.post1
Pygments==2.16.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-docx==0.8.11
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
regex==2023.8.8
requests==2.31.0
requests-file==1.5.1
requests-futures==1.0.1
requests-unixsocket==0.2.0
retrying==1.3.4
rich==13.5.2
S3Scanner==2.0.2
s3transfer==0.6.2
SecretStorage==2.3.1
service-identity==18.1.0
shodan==1.30.0
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sos==4.4
soupsieve==2.4.1
SQLAlchemy==2.0.20
sqlparse==0.4.4
ssh-import-id==5.10
subgpt==0.0.1
systemd-python==234
tenant-schemas-celery==2.2.0
termcolor==2.3.0
tldextract==3.4.4
tqdm==4.66.1
Twisted==18.9.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36
ujson==5.8.0
unattended-upgrades==0.1
untangle==1.2.1
urless @ file:///home/ubuntu/Tools/urless
urllib3==2.0.4
urlparse3==1.1
uuid==1.30
validators==0.21.2
vine==5.0.0
virtualenv==20.24.3
wadllib==1.3.3
wafw00f @ file:///home/ubuntu/Tools/wafw00f
waymore @ file:///home/ubuntu/Tools/waymore
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3
wrapt==1.15.0
XlsxWriter==3.1.2
xnLinkFinder @ file:///root/Tools/xnLinkFinder
yarl==1.9.2
zipp==3.16.2
zope.interface==6.0
